+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/ChatClient.java
@@ -182,8 +182,8 @@ public final class ChatClient {
      * Start receiving realtime notifications.
      * Until {@link ChatClient#stopRealtimeNotifications()} is called,
      * realtime notifications will be enabled and the corresponding registration will auto-renew.
-     * If there's error during registration initialization or renewal,
-     * realtime notification will be disabled and errorHandler will be called.
+     * If there's an error during registration initialization or renewal,
+     * realtime notifications will be disabled and {@code errorHandler} will be called.
      * @param context the Android app context
      * @param errorHandler error handler callback for registration failures
      * @throws RuntimeException if realtime notifications failed to start.
@@ -196,7 +196,7 @@ public final class ChatClient {
      * Start receiving realtime notifications.
      * Until {@link ChatClient#stopRealtimeNotifications()} is called,
      * realtime notifications will be enabled and the corresponding registration will auto-renew.
-     * If there's error during registration initialization or renewal, realtime notification will be disabled.
+     * If there's an error during registration initialization or renewal, realtime notifications will be disabled.
      * @param skypeUserToken the skype user token
      * @param context the Android app context
      * @throws RuntimeException if realtime notifications failed to start.
@@ -217,8 +217,8 @@ public final class ChatClient {
      * Register current device for receiving incoming push notifications via FCM.
      * Until {@link ChatClient#stopPushNotifications()} is called,
      * push notifications will be enabled and the corresponding registration will auto-renew.
-     * If there's error during registration initialization or renewal,
-     * push notifications will be disabled and errorHandler will be called.
+     * If there's an error during registration initialization or renewal,
+     * push notifications will be disabled and {@code errorHandler} will be called.
      * @param deviceRegistrationToken Device registration token obtained from the FCM SDK.
      * @param errorHandler error handler callback for registration failures
      * @throws RuntimeException if push notifications failed to start.
@@ -239,7 +239,7 @@ public final class ChatClient {
      * Handle incoming push notification.
      * Invoke corresponding chat event handle if registered.
      * @param pushNotification Incoming push notification payload from the FCM SDK.
-     * @throws IllegalStateException if push notifications has not started yet.
+     * @throws IllegalStateException if the push notification could not be handled.
      *
      * @return True if there's registered handler(s) for incoming push notification; otherwise, false.
      */