+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/ChatAsyncClient.java
@@ -288,7 +288,10 @@ public final class ChatAsyncClient {
      * @param skypeUserToken the skype user token
      * @param context the Android app context
      * @throws RuntimeException if realtime notifications failed to start.
+     *
+     * @deprecated Use {@link ChatAsyncClient#startRealtimeNotifications(Context, Consumer)}
      */
+    @Deprecated
     public void startRealtimeNotifications(String skypeUserToken, Context context) {
         if (this.signalingClient.hasStarted()) {
             return;